feat: expand site navigation with main page and logo
- Transform class-focused homepage into comprehensive navigation hub
- Create new MainPage with "Clumsy's World: Information Compendium" branding
- Add navigation menu for Spells, Items, Wiki, and Zones sections
- Move all original class functionality to dedicated Spells page
- Implement unique animated logo with magical theme and home navigation
- Fix ClassSpells back button to return to Spells instead of homepage
- Maintain complete data loading strategy and cache management
- Preserve all existing functionality: search, cart, hydration, caching
- Add responsive logo design with glassmorphism styling
- Ensure API endpoint consistency across all components
